pyproject.toml
[tool.poetry]
name = "transcription-service"
version = "0.1.0"
description = ""
authors = ["mmikolajczak <michal.mikolajczak@datarabbit.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
youtube-transcript-api = "^0.6.2"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
redis = "^5.0.8"
rq = "^1.16.2"
ffmpeg-python = "^0.2.0"
openai-whisper = "^20231117"
python-dotenv = "^1.0.1"
pytz = "^2024.1"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
